+++ firmware/controllers/trigger/trigger_emulator_algo.h
@@ -22,6 +22,5 @@ public:
 };
 
 void initTriggerEmulatorLogic(Logging *sharedLogger);
-void onConfigurationChangeTriggerCallback(engine_configuration_s *previousConfiguration);
 
 #endif /* TRIGGER_EMULATOR_ALGO_H_ */
